+++ node/src/main/scala/com/wavesplatform/utils/OptimisticLockable.scala
@@ -1,4 +1,4 @@
-package com.wavesplatform.events.repo
+package com.wavesplatform.utils
 
 import java.util.concurrent.locks.StampedLock
